+++ Cargo.toml
@@ -1,12 +1,15 @@
 [package]
 name = "rsr"
 version = "0.1.0"
-authors = ["Arthur Woimbée <arthur.woimbee@gmail.com>"]
+authors = ["Arthur Woimbée <arthur.woimbee@gmail.com>", "Mathieu LEDRU <matyo91@gmail.com>"]
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 clap = "2.33.0"
+reqwest = "0.9.24"
+serde_json = "1.0"
+mini-v8 = "0.1.0"
 regex = "1.3.1"
 rayon = "1.1"
