+++ docker-compose.yml
@@ -1,0 +1,11 @@
+version: "3.7"
+services:
+  uniflow-rust-client:
+    container_name: uniflow_rust_client
+    build: ./docker
+    volumes:
+      - .:/app
+    working_dir: /app
+    environment:
+      - CC=gcc-7
+      - CXX=g++-7
