+++ run
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+docker-compose run uniflow-rust-client cargo run "$@"
